Fixed namings and deleted unused api files
